feat(test): add broken link checker
Adds a broken link checker to the test suite.

This change introduces a script that starts a local server and crawls the site to find and validate all internal and external links, ensuring there are no 404 errors.

The following packages were added:
- broken-link-checker
- http-server
- start-server-and-test
- bhttp
- cheerio

A new 'test:links' script was added to package.json and integrated into the main 'npm test' command.